requirements.txt
requests
bs4
pytz
psutil
PILLOW

